@@ -4,7 +4,7 @@
   <groupId>org.esup-portail</groupId>
   <artifactId>esup-smsu-api-lib</artifactId>
   <packaging>jar</packaging>
-  <version>4.0.0</version>
+  <version>4.1.0-SNAPSHOT</version>
 
   <name>${project.groupId}:${project.artifactId}</name>
   <description>Provide web-services to send SMS. 
@@ -52,17 +52,17 @@
       <plugin>
 	<groupId>org.apache.maven.plugins</groupId>
 	<artifactId>maven-compiler-plugin</artifactId>
-	<version>2.3.2</version>
+	<version>3.11.0</version>
 	<configuration>
-	  <source>1.7</source>
-	  <target>1.7</target>
+	  <source>17</source>
+	  <target>17</target>
 	</configuration>
       </plugin>
 
       <plugin>
 	<groupId>org.apache.maven.plugins</groupId>
 	<artifactId>maven-source-plugin</artifactId>
-	<version>2.2.1</version>
+	<version>3.3.0</version>
 	<executions>
 	  <execution>
 	    <id>attach-sources</id>
@@ -76,7 +76,7 @@
       <plugin>
 	<groupId>org.sonatype.plugins</groupId>
 	<artifactId>nexus-staging-maven-plugin</artifactId>
-	<version>1.6.1</version>
+	<version>1.6.13</version>
 	<extensions>true</extensions>
 	<configuration>
 	  <serverId>ossrh</serverId>
@@ -91,7 +91,7 @@
 
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <spring-version>4.3.29.RELEASE</spring-version>
+    <spring-version>5.3.29</spring-version>
   </properties>
 
   <repositories>
@@ -108,7 +108,7 @@
 		<dependency>
 		  <groupId>commons-io</groupId>
 		  <artifactId>commons-io</artifactId>
-		  <version>[1.2, 1.99]</version>
+		  <version>2.11.0</version>
 		</dependency>
 
 		<dependency>
@@ -118,9 +118,9 @@
 		</dependency>
 
 		<dependency>
-			<groupId>org.codehaus.jackson</groupId>
-			<artifactId>jackson-mapper-asl</artifactId>
-			<version>1.9.11</version>
+			<groupId>com.fasterxml.jackson.core</groupId>
+			<artifactId>jackson-databind</artifactId>
+			<version>2.15.2</version>
 		</dependency>
 		
 		<dependency>
@@ -132,7 +132,7 @@
 		<dependency>
 		  <groupId>org.springframework</groupId>
   		  <artifactId>spring-beans</artifactId>
-  		  <version>4.3.29.RELEASE</version>
+  		  <version>${spring-version}</version>
 		</dependency>
 
   </dependencies>
