@@ -14,21 +14,21 @@
 
       <plugin>
         <artifactId>maven-war-plugin</artifactId>
-        <version>3.3.2</version>
+        <version>3.4.0</version>
       </plugin>
 
       <plugin>
 	<groupId>org.apache.maven.plugins</groupId>
 	<artifactId>maven-compiler-plugin</artifactId>
-	<version>2.3.2</version>
+	<version>3.11.0</version>
 	<configuration>
-	  <source>1.8</source>
-	  <target>1.8</target>
+	  <source>17</source>
+	  <target>17</target>
 	</configuration>
       </plugin>
 
       <plugin>
-	<groupId>org.mortbay.jetty</groupId>
+	<groupId>org.eclipse.jetty</groupId>
 	<artifactId>jetty-maven-plugin</artifactId>
 	<version>${jetty-version}</version>
       </plugin>
