@@ -1,6 +1,6 @@
 package org.esupportail.smsuapi.services.sms.impl.smsenvoi;
 
-import org.codehaus.jackson.annotate.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class SmsenvoiResponse {
