@@ -1,16 +1,13 @@
 package org.esupportail.smsuapi.services.client;
 
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.*;
 
 import org.esupportail.smsuapi.exceptions.InsufficientQuotaException;
 import org.esupportail.smsuapi.exceptions.InvalidParameterException;
 import org.esupportail.smsuapi.exceptions.UnknownMessageIdException;
 import org.esupportail.smsuapi.utils.HttpException;
 import org.esupportail.ws.remote.beans.TrackInfos;
-import org.junit.Test;
-import org.junit.internal.TextListener;
-import org.junit.runner.JUnitCore;
-import org.junit.runner.Result;
+import org.junit.jupiter.api.Test;
 
 public class HttpRequestSmsuapiWSTestOptional {
 
@@ -46,46 +43,46 @@ public class HttpRequestSmsuapiWSTestOptional {
 		return new HttpRequestSmsuapiWS(urlOk, usernameOk, passwordOk);
 	}
 	
-	@Test(expected = HttpException.Unreachable.class)  
-	public void testBadUrl() throws HttpException, UnknownMessageIdException {
-		wsBadUrl().messageInfos(4);
-	}
-	
-	@Test(expected = HttpException.Unreachable.class)  
-	public void testWSFirewallDrop() throws HttpException, UnknownMessageIdException {
-		wsFirewallDrop().messageInfos(4);
-	}
-	
-	@Test(expected = HttpException.Unreachable.class)  
-	public void testBadHost() throws HttpException, UnknownMessageIdException {
-		wsBadHost().messageInfos(4);
-	}
-	
-	@Test(expected = HttpException.Unreachable.class)  
-	public void testBadHttps() throws HttpException, UnknownMessageIdException {
-		wsBadHttps().messageInfos(4);
-	}
-	
-	@Test(expected = SmsuapiWS.AuthenticationFailedException.class)  
-	public void testBadUser() throws HttpException, UnknownMessageIdException {
-		wsBadUser().messageInfos(4);
-	}
-	
-	@Test(expected = UnknownMessageIdException.class)  
-	public void testMessageStatusUnknown() throws HttpException, UnknownMessageIdException {
-		wsOk().messageInfos(99999);
-	}
-	
 	@Test
-	public void testMessageStatus() throws HttpException, UnknownMessageIdException {
-		TrackInfos infos = wsOk().messageInfos(4);
-		assertEquals(infos.getNbDestTotal(), (Integer) 1);
-	}
-	
-	@Test(expected = InvalidParameterException.class)  
-	public void testSendMissingParam() throws HttpException, InsufficientQuotaException {
-		wsOk().sendSms(null, null, "test");
-	}
+    public void testBadUrl() {
+        assertThrows(HttpException.Unreachable.class, () -> wsBadUrl().messageInfos(4));
+    }
+
+    @Test
+    public void testWSFirewallDrop() {
+        assertThrows(HttpException.Unreachable.class, () -> wsFirewallDrop().messageInfos(4));
+    }
+
+    @Test
+    public void testBadHost() {
+        assertThrows(HttpException.Unreachable.class, () -> wsBadHost().messageInfos(4));
+    }
+
+    @Test
+    public void testBadHttps() {
+        assertThrows(HttpException.Unreachable.class, () -> wsBadHttps().messageInfos(4));
+    }
+
+    @Test
+    public void testBadUser() {
+        assertThrows(SmsuapiWS.AuthenticationFailedException.class, () -> wsBadUser().messageInfos(4));
+    }
+
+    @Test
+    public void testMessageStatusUnknown() {
+        assertThrows(UnknownMessageIdException.class, () -> wsOk().messageInfos(99999));
+    }
+
+    @Test
+    public void testMessageStatus() throws HttpException, UnknownMessageIdException {
+        TrackInfos infos = wsOk().messageInfos(4);
+        assertEquals((Integer) 1, infos.getNbDestTotal());
+    }
+
+    @Test
+    public void testSendMissingParam() {
+        assertThrows(InvalidParameterException.class, () -> wsOk().sendSms(null, null, "test"));
+    }
 	
 	/*
 	@Test(expected = InvalidParameterException.class)  
@@ -94,18 +91,18 @@ public class HttpRequestSmsuapiWSTestOptional {
 	}
 	*/
 	
-	@Test(expected = InsufficientQuotaException.class)
-	public void testCheckNoQuota() throws HttpException, InsufficientQuotaException {
-		wsNoQuota().mayCreateAccountCheckQuotaOk("", 2);
-	}
+    @Test
+    public void testCheckNoQuota() {
+        assertThrows(InsufficientQuotaException.class, () -> wsNoQuota().mayCreateAccountCheckQuotaOk("", 2));
+    }
 	
 	public void testCheckQuota() throws HttpException, InsufficientQuotaException {
 		wsOk().mayCreateAccountCheckQuotaOk("", 2);
 	}
 	
-	@Test(expected = InsufficientQuotaException.class)
-	public void testSendNoQuota() throws HttpException, InsufficientQuotaException {
-		wsNoQuota().sendSms(null, "0601010101", "test");
+	@Test
+	public void testSendNoQuota() {
+	    assertThrows(InsufficientQuotaException.class, () -> wsNoQuota().sendSms(null, "0601010101", "test"));
 	}
 	
 	@Test
@@ -117,12 +114,4 @@ public class HttpRequestSmsuapiWSTestOptional {
 		Integer id2_ = wsOk().sendSms(id2, "0601010101", "test");
 		assertEquals(id2, id2_);
 	}
-	
-	public static void all() {
-		JUnitCore runner = new JUnitCore();
-		runner.addListener(new TextListener(System.out));
-		Result result = runner.run(HttpRequestSmsuapiWSTestOptional.class);
-		System.out.println("tests finished: " + (result.wasSuccessful() ? "ok" : "KO"));
-	}
-
 }
